requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-appconf==1.0.2
django-ckeditor==5.6.1
django-cors-headers==3.0.1
django-filter==2.0.0
django-imagekit==4.0.2
django-jet==1.0.8
django-js-asset==1.1.0
django-maintenancemode==0.11.3
django-user-agents==0.4.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
IPy==1.0
jdcal==1.4.1
kombu==4.5.0
Markdown==3.0.1
mysqlclient==1.3.14
openpyxl==2.6.2
pilkit==2.0
Pillow==5.3.0
python-memcached==1.59
pytz==2018.7
PyYAML==5.1.2
redis==3.2.1
requests==2.21.0
six==1.12.0
ua-parser==0.8.0
urllib3==1.24.1
user-agents==2.0
vine==1.3.0
